Move base and panel classes into new location
