--- activity_browser/layouts/panels/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-from .bottom import BottomPanel
-from .left import LeftPanel
-from .panel import ABTab
-from .right import RightPanel
