--- activity_browser/layouts/panels/bottom.py
@@ -1,19 +1,0 @@
-from ..tabs import DebugTab
-from .panel import ABTab
-
-
-class BottomPanel(ABTab):
-    side = "bottom"
-
-    def __init__(self, *args):
-        super(BottomPanel, self).__init__(*args)
-
-        self.tabs = {"Debug": DebugTab(self)}
-        for tab_name, tab in self.tabs.items():
-            self.addTab(tab, tab_name)
-
-        self.setVisible(False)
-
-    def show_debug_window(self, toggle: bool = False):
-        self.setVisible(toggle)
-        self.tabs["Debug"].setVisible(toggle)
