+++ activity_browser/layouts/tabs/plugin.py
@@ -1,4 +1,4 @@
-from ..panels import ABTab
+from activity_browser.layouts.base import ABTab
 
 
 class PluginTab(ABTab):
@@ -9,7 +9,7 @@ class PluginTab(ABTab):
         :param plugin: instance of the plugin class
         :param panel: "left" or "right"
         """
-        super(PluginTab, self).__init__(parent)
+        super().__init__(parent)
         self.panel = panel
         self.plugin = plugin
         self.isPlugin = True
